FEM - 1677_1
+ Threshold of dvr will be in long instead of int.
+ isAutoPlay changed to boolean instead of Boolean.
+ Use metadata to detect the playbackType. In case there no metadata, as fallback use player to decide.
+ In live with dvr media always starts from position 0. So first events will contain dvr as playbackType. Avoid that by checking if impression was sent. Until then, report playbackType as live.

+++ kavaplugin/src/main/java/com/kaltura/playkit/plugins/kava/KavaAnalyticsConfig.java
@@ -33,7 +33,7 @@ public class KavaAnalyticsConfig {
     private String playbackContext;
     private String referrerAsBase64;
     private String baseUrl = DEFAULT_BASE_URL;
-    private int  dvrThreshold = Consts.DISTANCE_FROM_LIVE_THRESHOLD;
+    private long  dvrThreshold = Consts.DISTANCE_FROM_LIVE_THRESHOLD;
     private String customVar1, customVar2, customVar3;
 
     public KavaAnalyticsConfig setUiConfId(int uiConfId) {
@@ -108,7 +108,7 @@ public class KavaAnalyticsConfig {
         return baseUrl;
     }
 
-    int getDvrThreshold() {
+    long getDvrThreshold() {
         return dvrThreshold;
     }
